refactor(api): create api directory for FastAPI module
