+++ api/quizzify/routers/artists/__init__.py
