+++ api/quizzify/routers/artists/service.py
