+++ api/quizzify/routers/auth/service.py
