+++ api/quizzify/utils/singleton.py
