+++ docker-compose.yaml
@@ -5,7 +5,7 @@ services:
   quizzify-api:
     build:
       context: .
-      dockerfile: Dockerfile
+      dockerfile: api/Dockerfile
     container_name: quizzify-api
     command: uvicorn quizzify.main:app --host 0.0.0.0 --port 8000 --log-config quizzify/logging.conf
     restart: always
@@ -33,7 +33,7 @@ services:
       - .env
     volumes:
       # Mount the volume to persist the data
-      - ./quizzify/databases/init.sql:/docker-entrypoint-initdb.d/init.sql
+      - ./db/init.sql:/docker-entrypoint-initdb.d/init.sql
       # Mount the init script to create the database as an entrypoint to the container
       - postgres_data:/var/lib/postgresql/data
     networks:
