+++ tests/quizzify/utils/test_helpers.py
@@ -1,4 +1,4 @@
-from quizzify.utils.helpers import encode_str_to_base64
+from app.utils.helpers import encode_str_to_base64
 
 
 def test_encode_str_to_base64():
